% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics_identification.R
\name{add_penalty_statistics}
\alias{add_penalty_statistics}
\title{add_penalty_statistics}
\usage{
add_penalty_statistics(match_t, minimum_matching_mutations)
}
\arguments{
\item{match_t}{object that contains the matching variants}

\item{minimum_matching_mutations}{a numerical giving the minimum amount of
mutations that has to match between query and training sample for a 
positive prediction}
}
\value{
The updated statistics
}
\description{
Add penalty statistics to results
}
