% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parse_COSMIC_CCLE_CCLs.R
\name{parse_cosmic_genotype_data}
\alias{parse_cosmic_genotype_data}
\title{parse_cosmic_genotype_data}
\usage{
parse_cosmic_genotype_data(cosmic_file, ref_gen = "GRCH37")
}
\arguments{
\item{cosmic_file}{Path to cosmic clp file in hard disk}

\item{ref_gen}{Reference genome version}
}
\value{
The R Table sim_list which contains the CoSMIC CLP fingerprints
}
\description{
Parses cosmic genotype data
}
