% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Show_database_information.R
\name{read_library_names}
\alias{read_library_names}
\title{Library Name Reader}
\usage{
read_library_names(ref_gen)
}
\arguments{
\item{ref_gen}{a character vector specifying the reference genome 
version. All training sets are associated with a reference genome 
version. Default is \code{"GRCH37"}.}
}
\value{
Returns a character vector of the contained libraries
}
\description{
This function procides information on the reference library names
}
\examples{
read_library_names(ref_gen = "GRCH37")
}
